certomancer-service-docker/requirements.txt
certomancer==0.3.0-dev1
redis>=3.5.3
Werkzeug>=1.0.1
Jinja2>=2.11.3
PyYAML>=5.4.1
asn1crypto>=1.4.0